requirements.txt
streamlit
openai
PyPDF2
pandas
openpyxl
python-dotenv
